Lazily update description file size to avoid API calls
Initially, I tried to simply leave out GetAttr, and then tried always
returning a size of 0. The thought was, this is what a /proc file looks
like. This doesn't work. Even though I can see the READ request
returning many bytes, the read() system call returns 0.

From what I can tell, it is the kernel that prevents this from working
as I'd hoped. During a read(), fuse_file_read_iter() may call
fuse_update_attributes() which, if we provide the right size at that
time, will cause the read() to return the correct number of bytes.

So, we wait until after the file is opened, and we've already made the
API call to set this node's size. Even if it changes, we still always
update the node size immediately.