Call file.GetAttr() in each node GetAttr() if file is given
This avoids the need to make API calls from GetAttr() which is called
often. Instead, we allow GetAttr() to return a size of 0 when there is
no open file object (e.g. stat()), but return our in-memory file's size
when GetAttr() is called with a file object (e.g. fstat() or read()).

This ensures that the kernel knows the correct file size during a
read(), and allows us to implement a file with 0-sized /proc file
semantics.