Disable timeouts in FUSE options
attr_timeout is

> The timeout in seconds for which file/directory attributes are cached.
> The default is 1.0 second.

With a nonzero attr_timeout, the results of GetAttr() calls are cached
in the kernel.

This is not compatible with my GetAttr() scheme (abb48d25fc), where
stat() returns a size of zero (avoiding an API call), but read() still
returns data as expected. This is because the result of first
(file == nil) GetAttr() call is cached, the subsequent call
(file != nil) never happens.